+++ plugin-src/translators/shapeWithText/index.ts
@@ -1,4 +1,3 @@
-export * from './exportShapeWithTextSvg';
 export * from './extractTextLayout';
 export * from './translateShapeWithTextContent';
 export * from './translateShapeWithTextGeometry';
