+++ plugin-src/translators/shapeWithText/parseSvgAttrs.ts
@@ -2,6 +2,11 @@
 // Returns the `name="value"` pairs from inside an SVG element's opening tag.
 
 const ATTR_REGEX = /([\w-]+)\s*=\s*"([^"]*)"/g;
+const DEFS_REGEX = /<defs\b[^>]*>[\s\S]*?<\/defs>/gi;
+
+// Drop <defs> so clipPath / mask / gradient content isn't mistaken for the
+// visible geometry or text.
+export const stripSvgDefs = (svg: string): string => svg.replace(DEFS_REGEX, '');
 
 export const parseSvgAttrs = (input: string): Record<string, string> => {
   const result: Record<string, string> = {};
