+++ plugin-src/translators/vectors/index.ts
@@ -1,3 +1,5 @@
+export * from './applyMatrixToCommand';
+export * from './normalizeCommands';
 export * from './translateCommands';
 export * from './translateNonRotatedCommands';
 export * from './translateRotatedCommands';
