+++ plugin-src/translators/vectors/normalizeCommands.ts
