+++ plugin-src/translators/vectors/translateNonRotatedCommands.ts
@@ -1,7 +1,5 @@
 import type { Command } from 'svg-path-parser';
 
-import { applyMatrixToPoint } from '@plugin/utils';
-
 export const translateNonRotatedCommands = (
   commands: Command[],
   baseX: number = 0,
@@ -50,27 +48,6 @@ export const applyBase = (command: Command, baseX: number, baseY: number): Comma
         x: command.x + baseX,
         y: command.y + baseY
       };
-    default:
-      return command;
-  }
-};
-
-export const applyMatrixToCommand = (command: Command, matrix: Transform): Command => {
-  const project = (x: number, y: number): { x: number; y: number } => {
-    const [px, py] = applyMatrixToPoint(matrix, [x, y]);
-    return { x: px, y: py };
-  };
-
-  switch (command.command) {
-    case 'lineto':
-    case 'moveto':
-      return { ...command, ...project(command.x, command.y) };
-    case 'curveto': {
-      const c1 = project(command.x1, command.y1);
-      const c2 = project(command.x2, command.y2);
-      const end = project(command.x, command.y);
-      return { ...command, x1: c1.x, y1: c1.y, x2: c2.x, y2: c2.y, x: end.x, y: end.y };
-    }
     default:
       return command;
   }
